requirements.txt
streamlit>=1.30.0
tensorflow>=2.13.0
keras>=2.13.0
opencv-python-headless>=4.8.0
scikit-learn>=1.4.2
pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.25.0
pandas>=2.2.2
seaborn>=0.13.2
joblib>=1.4.2
matplotlib>=3.8.0
numba>=0.59.0
